Добавлено тести

+++ src/main/java/ua/lviv/lgs/admissionsOffice/controller/ApplicationController.java
@@ -67,7 +67,7 @@ public class ApplicationController {
 	
 	@PreAuthorize("hasAuthority('USER')")
 	@PostMapping("/create")
-	public String createApplication(@RequestParam Map<String, String> form,	@RequestParam("files") MultipartFile[] supportingDocuments,
+	public String createApplication(@RequestParam Map<String, String> form,	@RequestParam("supportingDocument") MultipartFile[] supportingDocuments,
 			@Valid Application application, BindingResult bindingResult, Model model) throws IOException {
 		Map<String, String> znoMarksErrors = applicationService.getZnoMarksErrors(form);
 		Map<String, String> supportingDocumentErrors = supportingDocumentService.getSupportingDocumentErrors(supportingDocuments);
@@ -77,7 +77,8 @@ public class ApplicationController {
             model.mergeAttributes(errors);
             model.mergeAttributes(supportingDocumentErrors);
             model.addAttribute(!znoMarksErrors.isEmpty() ? "znoMarksErrorMessage" : "", "При заповненні балів з ЗНО були виявлені помилки: " +
-            		znoMarksErrors.values() + ". Спробуйте заповнити форму ще раз!");
+            		znoMarksErrors.values() + ". \n"
+            				+ "Спробуйте заповнити форму ще раз!");
             model.addAttribute(form.get("speciality").isEmpty() ? "specialityError" : "", "Поле Спеціальність не може бути порожньою!");
     		model.addAttribute("specialities", specialityService.findByRecruitmentCompletedFalse());
     		
@@ -97,7 +98,14 @@ public class ApplicationController {
 	}
 	
 	@GetMapping("/edit")
-	public String viewEditForm(@RequestParam("id") Application application, Model model) {
+	public String viewEditForm(@RequestParam("id") Application application, HttpSession session, Model model) {
+		User currentUser = ((User) session.getAttribute("user"));
+		if (currentUser.getAccessLevels().contains(AccessLevel.valueOf("USER"))
+				&& !application.getApplicant().getId().equals(currentUser.getId())
+				|| application.getRatingList().isAccepted()) {
+			return "redirect:/403";
+		}
+		
 		model.addAttribute("aplication", application);
 		model.addAttribute("specialities", specialityService.findByRecruitmentCompletedFalse());
 		model.addAttribute("downloadURI", ServletUriComponentsBuilder.fromCurrentContextPath().path("/downloadFile/").toUriString());
@@ -107,7 +115,7 @@ public class ApplicationController {
 
 	@PostMapping("/edit")
 	public String updateApplication(@RequestParam("id") Application application, @RequestParam Map<String, String> form,
-			@RequestParam("files") MultipartFile[] supportingDocuments, HttpSession session, @Valid Application updatedApplication,
+			@RequestParam("supportingDocument") MultipartFile[] supportingDocuments, HttpSession session, @Valid Application updatedApplication,
 			BindingResult bindingResult, Model model) throws IOException {
 		Map<String, String> znoMarksErrors = applicationService.getZnoMarksErrors(form);
 		Map<String, String> supportingDocumentErrors = supportingDocumentService.getSupportingDocumentErrors(supportingDocuments);
@@ -124,8 +132,17 @@ public class ApplicationController {
 			return "applicationEditor";
 		}
 		
-		applicationService.updateApplication(updatedApplication, form, supportingDocuments);
+		boolean applicationExists = !applicationService.updateApplication(updatedApplication, form, supportingDocuments);
+		
+		if (applicationExists) {
+			model.addAttribute("applicationExistsMessage", "На обрану спеціальність заявка вже існує!");
+			model.addAttribute("aplication", application);
+			model.addAttribute("specialities", specialityService.findByRecruitmentCompletedFalse());
+			model.addAttribute("downloadURI", ServletUriComponentsBuilder.fromCurrentContextPath().path("/downloadFile/").toUriString());
 
+			return "applicationEditor";
+		}
+		
 		if (((User) session.getAttribute("user")).getAccessLevels().contains(AccessLevel.valueOf("ADMIN"))) {
 			return "redirect:/application/notAcceptedApps";
 		}
@@ -135,7 +152,14 @@ public class ApplicationController {
 	
 	@PreAuthorize("hasAuthority('USER')")
 	@GetMapping("/delete")
-	public String deleteApplication(@RequestParam("id") Application application) {
+	public String deleteApplication(@RequestParam("id") Application application, HttpSession session) {
+		User currentUser = ((User) session.getAttribute("user"));
+		if (currentUser.getAccessLevels().contains(AccessLevel.valueOf("USER"))
+				&& !application.getApplicant().getId().equals(currentUser.getId())
+				|| application.getSpeciality().isRecruitmentCompleted()) {
+			return "redirect:/403";
+		}
+		
 		applicationService.deleteApplication(application);
 
 		return "redirect:/application";
