+++ src/main/java/ua/lviv/lgs/admissionsOffice/controller/FacultyController.java
@@ -2,6 +2,7 @@ package ua.lviv.lgs.admissionsOffice.controller;
 
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.List;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
@@ -32,14 +33,17 @@ public class FacultyController {
 	
 	@GetMapping
 	public String viewFacultyList(Model model) {
-		model.addAttribute("faculties", facultyService.findAll());
-		model.addAttribute("aplicationsByFaculty", facultyService.countApplicationsByFaculty());
+		List<Faculty> facultiesList = facultyService.findAll();
+		Map<Faculty, Integer> countApplicationsByFaculty = facultyService.countApplicationsByFaculty();
+		model.addAttribute("faculties", facultiesList);
+		model.addAttribute("aplicationsByFaculty", countApplicationsByFaculty);
 
 		return "facultyList";
 	}
 	
 	@GetMapping("/create")
-	public String viewCreationForm(@RequestParam(name = "superRefererURI", required = false) String superRefererURI, HttpServletRequest request, Model model) throws URISyntaxException {
+	public String viewCreationForm(@RequestParam(name = "superRefererURI", required = false) String superRefererURI,
+			HttpServletRequest request, Model model) throws URISyntaxException {
 		model.addAttribute("subjects", subjectService.findAll());
 		model.addAttribute("refererURI", new URI(request.getHeader("referer")).getPath());
 
@@ -88,6 +92,11 @@ public class FacultyController {
 	
 	@GetMapping("/edit")
 	public String viewEditForm(@RequestParam("id") Faculty faculty, Model model) {
+		Map<Faculty, Integer> countApplicationsByFaculty = facultyService.countApplicationsByFaculty();
+		if (countApplicationsByFaculty.get(faculty) != 0) {
+			return "redirect:/403";	
+		}
+		
 		model.addAttribute("faculty", faculty);
 		model.addAttribute("subjects", subjectService.findAll());
 		
@@ -95,7 +104,8 @@ public class FacultyController {
 	}
 
 	@PostMapping("/edit")
-	public String updateFaculty(@RequestParam("id") Faculty faculty, @RequestParam Map<String, String> form, @Valid Faculty updatedFaculty, BindingResult bindingResult, Model model) {
+	public String updateFaculty(@RequestParam("id") Faculty faculty, @RequestParam Map<String, String> form,
+			@Valid Faculty updatedFaculty, BindingResult bindingResult, Model model) {
 		if (bindingResult.hasErrors()) {
 			Map<String, String> errors = ControllerUtils.getErrors(bindingResult);
 			model.mergeAttributes(errors);
@@ -105,13 +115,25 @@ public class FacultyController {
 			return "facultyEditor";
 		}
 
-		facultyService.updateFaculty(updatedFaculty, form);
-
+		boolean facultyExists = !facultyService.updateFaculty(updatedFaculty, form);
+		
+		if (facultyExists) {
+			model.addAttribute("facultyExistsMessage", "Такий факультет вже існує!");
+			model.addAttribute("faculty", faculty);
+			model.addAttribute("subjects", subjectService.findAll());
+						
+			return "facultyEditor";
+		}
+		
 		return "redirect:/faculty";
 	}
 	
 	@GetMapping("/delete")
 	public String deleteFaculty(@RequestParam("id") Faculty faculty) {
+		if (!faculty.getExamSubjects().isEmpty() && !faculty.getSpecialities().isEmpty()) {
+			return "redirect:/403";	
+		}
+		
 		facultyService.deleteFaculty(faculty);
 
 		return "redirect:/faculty";
